using ruby-prof Flat mode for Array#select optimisation
